Crate net_ensembles

Libary for my simulations Here I will define the node structs etc.

You probably want to take a look at the struct Graph from the module graph.

Also, take a look at the module er if you want to do something with Erdős-Rényi networks

Example

use net_ensembles;
use net_ensembles::Node;

Re-exports

pub use node::Node;

pub use er::ER;

pub use er::ErStep;

pub use graph::Graph;

pub use node::TestNode;

pub use graph::DEFAULT_DOT_OPTIONS;

Modules

er
Create Erdős-Rényi networks

graph
Topology

node
Trait needed for ER graphs etc.