Module net_ensembles::dual_graph

Structs

Adj

BfsDualIndex

DfsDualContained

DfsDualIndex

DualGraph

Enums

AddEdgeError

AdjSingle

WhichGraph

Stores information T and the graph the information corresponds to

Traits

AdjTrait

Type Definitions

DefaultSDG

DualIndex

Index which also stores for which graph the index is

MultiDualGraph

SingleDualGraph