Struct net_ensembles::er::ER

pub struct ER<T: Node, R: Rng + SeedableRng> { /* fields omitted */ }

Implements Erdős-Rényi graph

Methods

impl<T: Node, R: Rng + SeedableRng> ER<T, R>

pub fn new(n: u32, c_target: f64, rng: R) -> Self

Initialize

create new ER graph with n vertices

target connectivity c_target

rng is consumed and used as random number generator in the following

internally uses Graph<T>::new(n)

generates random edges according to ER probability (see ER::randomize)

pub fn randomize(&mut self)

Randomizes the edges according to Er probabilities

this is used by ER::new to create the initial topology

you can use this for sampling the ensemble

runs in O(vertices * vertices)

pub fn target_connectivity(&self) -> f64

returns target connectivity

Explanation

The target connectivity c_target is used to calculate the probability p, that any two vertices i and j (where i != j) are connected.

p = c_target / (N - 1) where N is the number of vertices in the graph

pub fn random_step(&mut self) -> ErStep

Monte Carlo steps

use this to perform a Monte Carlo step

result ErStep can be used to undo the step with self.undo_step(result)

pub fn random_steps(&mut self, count: usize) -> Vec<ErStep>

Monte Carlo steps

use this to perform multiple Monte Carlo steps at once

result Vec<ErStep> can be used to undo the steps with self.undo_steps(result)

pub fn undo_step(&mut self, step: ErStep) -> Result<(), GraphErrors>

Undo a Monte Carlo step

adds removed edge, or removes added edge, or does nothing

if it returns an Err value, you probably used the function wrong

Important:

Restored graph is the same as before the random step except the order of nodes in the adjacency list might be shuffled!

pub fn undo_steps(&mut self, steps: Vec<ErStep>) -> Result<(), GraphErrors>

Undo a Monte Carlo step

adds removed edges, removes added edge etc. in the correct order

if it returns an Err value, you probably used the function wrong

Important:

Restored graph is the same as before the random steps except the order of nodes in the adjacency list might be shuffled!

pub fn sort_adj(&mut self)

Sorting adjecency lists

calls sort_unstable() on all adjecency lists

pub fn graph(&self) -> &Graph<T>

returns reference to the underlying topology aka, the Graph<T>

Use this to call functions regarding the topology

pub fn rng(&mut self) -> &mut R

Access RNG

If, for some reason, you want access to the internal random number generator: Here you go