Struct net_ensembles::er_c::ErEnsembleC

pub struct ErEnsembleC<T, R>where
    T: Node,{ /* private fields */ }

Implements Erdős-Rényi graph ensemble

variable number of edges

targets a connectivity

Sampling

for simple sampling look at SimpleSample trait

for markov steps look at MarkovChain trait

Other

for topology functions look at GenericGraph

to access underlying topology or manipulate additional data look at WithGraph trait

to use or swap the random number generator, look at HasRng trait

Implementations

impl<T, R> ErEnsembleC<T, R>where
T: Node + SerdeStateConform,
R: Rng,

pub fn new(n: usize, c_target: f64, rng: R) -> Self

Initialize

create new ErEnsembleC with:

n vertices

target connectivity c_target

rng is consumed and used as random number generator in the following

internally uses Graph<T>::new(n)

generates random edges according to ER model

pub fn make_connected(&mut self)

Experimental! Connect the connected components

adds edges, to connect the connected components

panics if no vertices are in the graph

intended as starting point for a markov chain, if you require connected graphs

do not use this to independently (simple-) sample connected networks, as this will skew the statistics

This is still experimental, this member might change the internal functionallity resulting in different connected networks, without prior notice

This member might be removed in braking releases

pub fn target_connectivity(&self) -> f64

returns target connectivity

Explanation

The target connectivity c_target is used to calculate the probability p, that any two vertices i and j (where i != j) are connected.

p = c_target / (N - 1) where N is the number of vertices in the graph

pub fn set_target_connectivity(&mut self, c_target: f64)

set new value for target connectivity

Note

will only set the value (and probability), which will be used from now on

if you also want to create a new sample, call randomize afterwards