Module net_ensembles::er_m

Erdős-Rényi with constant number of edges

Draw from an Erdős-Rényi graph ensemble

In this model, all possible edges are equally likely

The number of edges is fixed

Citations

P. Erdős and A. Rényi, “On the evolution of random graphs,” Publ. Math. Inst. Hungar. Acad. Sci. 5, 17-61 (1960)

Structs

ErEnsembleM

Implements Erdős-Rényi graph ensemble

ErStepM

Storing the information about which edges were deleted or added