Struct net_ensembles::graph::Graph

pub struct Graph<T: Node> { /* fields omitted */ }

Contains the topology and implements functions for analyzing topology

used for graph ensembles

Example:

A graph, where each node stores a phase

use net_ensembles::{Node,Graph};
use std::fs::File;
use std::io::prelude::*;
// define your own vertices, if you need to store extra information at each vertex
#[derive(Debug, Clone)]
pub struct PhaseNode {phase: f64,}

// implement whatever you need
impl PhaseNode {
    pub fn set_phase(&mut self, phase: f64) {
        self.phase = phase;
    }

    pub fn get_phase(&self) -> f64 {
        self.phase
    }
}

// implement the trait `Node`
impl Node for PhaseNode {
    fn new_from_index(index: u32) -> Self {
        PhaseNode { phase: index as f64 * 10.0}
    }
}

// now you can create an empty graph
let mut graph: Graph<PhaseNode> = Graph::new(4);
for i in 0..4 {
    assert_eq!(
      graph.at(i).get_phase(),
      i as f64 * 10.0
    );
}

// and fill it with edges
for i in 0..4 {
    graph.add_edge(i, (i + 1) % 4).unwrap();
}

// you can manipulate the extra information stored at each Vertex
for i in 0..4 {
    graph.at_mut(i).set_phase(i as f64 * 0.5);
}

// you can, of course, also access the information
for i in 0..4 {
    assert_eq!(
        graph.at(i).get_phase(),
        i as f64 * 0.5
    );
}

// if you want to visualize your graph, you can generate a string with graphviz representation
let dot = graph.to_dot_with_labels(
    "",
    |contained, index|
           format!(
                "Phase: {} at index {}",
                contained.get_phase(),
                index
            )
);
// which you can store in a dot file
let mut f = File::create("phase_example.dot").expect("Unable to create file");
f.write_all(dot.as_bytes()).expect("Unable to write data");

// or just print it out
println!("{}", dot);

phase_example.dot will then contain

graph G{

	0 1 2 3 ;
	"0" [label="Phase: 0 at index 0"];
	"1" [label="Phase: 0.5 at index 1"];
	"2" [label="Phase: 1 at index 2"];
	"3" [label="Phase: 1.5 at index 3"];
	0 -- 1
	0 -- 3
	1 -- 2
	2 -- 3
}

Now you can use circo or similar programs to create a pdf from that. Search for graphviz for more info.

Example 2

if you also want to be able to store your graph, you have to override two functions of the Node trait, as shown below

use net_ensembles::{Node,Graph};
use std::fs::File;
use std::io::prelude::*;
// define your own vertices, if you need to store extra information at each vertex
#[derive(Debug, Clone)]
pub struct PhaseNode {phase: f64,}

impl PhaseNode {
    pub fn set_phase(&mut self, phase: f64) {
        self.phase = phase;
    }

    pub fn get_phase(&self) -> f64 {
        self.phase
    }
}

impl Node for PhaseNode {
    fn new_from_index(index: u32) -> Self {
        PhaseNode { phase: 10.0 * index as f64 }
    }

    // Override this, to save the graph
    fn make_string(&self) -> Option<String> {
        Some(format!("phase: {},", self.phase))
    }

    // Override this, since you want to load the stored graph
    fn parse_str(to_parse: &str) -> Option<(&str, Self)>
        where Self: Sized
    {
        let identifier = "phase: ";
        // searching for identifier
        let mut split_index = to_parse.find(identifier)?;

        // add length of identifier to index
        split_index += identifier.len();

        // new string slice to skip to our identifier
        let remaining_to_parse = &to_parse[split_index..];

        // find out, where our data ends
        split_index = remaining_to_parse.find(",")?;

        // create new string slice, beginning after what Node::make_string() created
        let (phase_str, mut remaining_to_parse) = remaining_to_parse.split_at(split_index);
        remaining_to_parse = &remaining_to_parse[1..];

        // parse our data
        let phase = phase_str.parse::<f64>().ok()?;
        let node = PhaseNode{ phase };

        // return our struct as option
        Some((remaining_to_parse, node))
    }
}

// now you can create an empty graph
let mut graph: Graph<PhaseNode> = Graph::new(4);
for i in 0..4 {
    assert_eq!(
      graph.at(i).get_phase(),
      i as f64 * 10.0
    );
}

// and fill it with edges
for i in 0..4 {
    graph.add_edge(i, (i + 1) % 4).unwrap();
}

// you can manipulate the extra information stored at each Vertex
for i in 0..4 {
    graph.at_mut(i).set_phase(i as f64 * 0.5);
}

// you can, of course, also access the information
for i in 0..4 {
    assert_eq!(
        graph.at(i).get_phase(),
        i as f64 * 0.5
    );
}

// if you want to visualize your graph, you can generate a string with graphviz representation
let dot = graph.to_dot_with_labels(
    "",
    |contained, index|
           format!(
                "Phase: {} at index {}",
                contained.get_phase(),
                index
            )
);
// which you can store in a dot file
let mut f = File::create("phase_example_still_works.dot").expect("Unable to create file");
f.write_all(dot.as_bytes()).expect("Unable to write data");

// or just print it out
println!("{}", dot);

// if you want to store your graph, you can do the following:
// first get string representation
let s = graph.to_string();
// open file
let mut graph_file = File::create("store_graph_example.dat").expect("Unable to create file");
// write to file
graph_file.write_all(s.as_bytes()).expect("Unable to write data");
// close file
drop(graph_file);

// now, if you want to load your network:
let mut read_in = File::open("store_graph_example.dat").expect("unable to open file");
let mut test_data = String::new();
read_in.read_to_string(&mut test_data).expect("unable to read file");

// now we read the string. We still have to parse it:
let (_, graph2) = Graph::<PhaseNode>::parse_str(&test_data).unwrap();

// now, to show, that the graphs are equal, here is one of my test functions:
// modified for this example, which is a doc-test, so this example also serves as unit test

fn assert_equal_graphs(g1: &Graph<PhaseNode>, g2: &Graph<PhaseNode>) {
    assert_eq!(g1.edge_count(), g2.edge_count());
    assert_eq!(g1.vertex_count(), g2.vertex_count());
    for (n0, n1) in g2.container_iter().zip(g1.container_iter()) {
        assert_eq!(n1.id(), n0.id());
        assert_eq!(n0.neighbor_count(), n1.neighbor_count());

        for (i, j) in n1.neighbors().zip(n0.neighbors()) {
            assert_eq!(i, j);
        }
    }

    for i in 0..g2.vertex_count() as usize {
        assert_eq!(
            g1.at(i).get_phase(),
            g2.at(i).get_phase()
        );
    }
}
// lets use it
assert_equal_graphs(&graph, &graph2);

// you can also clone the graph, if you need:
let clone = graph.clone();
assert_equal_graphs(&graph, &clone);
let clone2 = graph2.clone();
assert_equal_graphs(&clone, &clone2);

Methods

impl<T: Node> Graph<T>

pub fn new(size: u32) -> Self

Create new graph with size nodes and no edges

pub fn clear_edges(&mut self)

removes all edges from the graph

inexpensive O(1), if there are no edges to begin with

O(vertices) otherwise

pub fn sort_adj(&mut self)

Sort adjecency lists

If you depend on the order of the adjecency lists, you can sort them

keep in mind, that this can be costly

pub fn parse_str(to_parse: &str) -> Option<(&str, Self)>

parse from str

tries to parse Graph from a str.

will ignore leading whitespaces and other chars, as long as they do not match "next_id: "

returns None if failed

Return

returns string slice beginning directly after the part, that was used to parse

the Graph resulting form the parsing

pub fn container(&self, index: usize) -> &NodeContainer<T>

get NodeContainer at index

use this to iterate over neighbors

use this to check, if vertices are adjacent

Warning

panics if index out of bounds

pub fn container_iter(&self) -> Iter<NodeContainer<T>>

get iterator over NodeContainer in order of the indices

pub fn at(&self, index: usize) -> &T

For your calculations etc.

read access to your struct T, stored at each vertex, that implements Node trait

see first code example (beginning of this page)

pub fn at_mut(&mut self, index: usize) -> &mut T

For your calculations etc.

write access to your struct T, stored at each vertex, that implements Node trait

see first code example (beginning of this page)

pub fn vertex_count(&self) -> u32

returns number of vertices present in graph

pub fn average_degree(&self) -> f32

pub fn add_edge(&mut self, index1: u32, index2: u32) -> Result<(), GraphErrors>

Adds edge between nodes index1 and index2

ErrorCases:

Error	Reason

GraphErrors::IndexOutOfRange	index1 or index2 larger than self.vertex_count()

GraphErrors::IdenticalIndices	index2 == index1 not allowed!

GraphErrors::EdgeExists	requested edge already exists!

pub fn remove_edge(
&mut self,
index1: u32,
index2: u32
) -> Result<(), GraphErrors>

Removes edge between nodes index1 and index2

ErrorCases:

Error	Reason

GraphErrors::IndexOutOfRange	index1 or index2 larger than self.vertex_count()

GraphErrors::IdenticalIndices	index2 == index1 not allowed!

GraphErrors::EdgeDoesNotExist	requested edge does not exists

pub fn edge_count(&self) -> u32

returns total number of edges in graph

pub fn edge_count_at(&self, index: usize) -> Option<usize>

Important traits for Dfs<'a, T>
impl<'a, T> Iterator for Dfs<'a, T> where
T: 'a + Node, type Item = &'a T;

pub fn dfs(&self, index: u32) -> Dfs<T>

returns Iterator

the iterator will iterate over the vertices in depth first search order, beginning with vertex index.

iterator returns node

Order

Order is guaranteed to be in DFS order, however if this order is not unambigouse adding edges and especially removing edges will shuffle the order.

Note:

Will only iterate over vertices within the connected component that contains vertex index

Important traits for DfsWithIndex<'a, T>
impl<'a, T> Iterator for DfsWithIndex<'a, T> where
T: 'a + Node, type Item = (u32, &'a T);

pub fn dfs_with_index(&self, index: u32) -> DfsWithIndex<T>

returns Iterator

the iterator will iterate over the vertices in depth first search order, beginning with vertex index.

Iterator returns tuple (index, node)

Order

Order is guaranteed to be in DFS order, however if this order is not unambigouse adding edges and especially removing edges will shuffle the order.

Note:

Will only iterate over vertices within the connected component that contains vertex index

Important traits for Bfs<'a, T>
impl<'a, T> Iterator for Bfs<'a, T> where
T: 'a + Node, type Item = (u32, &'a T, u32);

pub fn bfs_index_depth(&self, index: u32) -> Bfs<T>

returns Iterator

the iterator will iterate over the vertices in breadth first search order, beginning with vertex index.

Iterator returns tuple (index, node, depth)

depth

starts at 0 (i.e. the first element in the iterator will have depth = 0)

depth equals number of edges in the shortest path from the current vertex to the first vertex (i.e. to the vertex with index index)

Order

Order is guaranteed to be in BFS order, however if this order is not unambigouse adding edges and especially removing edges will shuffle the order.

Note:

Will only iterate over vertices within the connected component that contains vertex index

pub fn is_connected(&self) -> Option<bool>

result	condition

None	if graph does not contain any vertices

Some(true)	else if all vertices are connected by paths of edges

Some(false)	otherwise

pub fn q_core(&self, q: u32) -> Option<u32>

definition

Calculates the size of the q-core (i.e. number of nodes in the biggest possible set of nodes, where all nodes from the set are connected with at least q other nodes from the set)

returns None if impossible to calculate (e.g. vertex_count == 0 or q <= 1)

Example

use net_ensembles::TestNode;
use net_ensembles::Graph;

let graph: Graph<TestNode> = Graph::new(0);
assert_eq!(graph.q_core(1), None);
assert_eq!(graph.q_core(2), None);

let graph2: Graph<TestNode> = Graph::new(1);

assert_eq!(graph2.q_core(1), None);
assert_eq!(graph2.q_core(2), Some(0));

// create complete graph
let mut graph3: Graph<TestNode> = Graph::new(20);
for i in 0..graph3.vertex_count() {
    for j in i+1..graph3.vertex_count() {
        graph3.add_edge(i, j).unwrap();
    }
}

// since this is a complete graph, the q-core should always consist of 20 nodes
// as long as q < 20, as every node has 19 neighbors
for i in 2..20 {
    assert_eq!(graph3.q_core(i), Some(20));
}
assert_eq!(graph3.q_core(20), Some(0));

pub fn connected_components(&self) -> Vec<u32>

compute sizes of all connected components

the number of connected components is the size of the returned vector, i.e. result.len()

returns empty vector, if graph does not contain vertices

returns (reverse) ordered vector of sizes of the connected components, i.e. the biggest component is of size result[0] and the smallest is of size result[result.len() - 1]

pub fn leaf_count(&self) -> usize

Count number of leaves in the graph, i.e. vertices with exactly one neighbor

pub fn to_dot(&self) -> String

Creates String which contains the topology of the network in a format that can be used by circo etc. to generate a pdf of the graph.

indices are used as labels

search for graphviz to learn about .dot format

pub fn to_dot_with_labels<F>(&self, dot_options: &str, f: F) -> String where
F: Fn(&T, usize) -> String,

Example

use std::fs::File;
use std::io::prelude::*;
use net_ensembles::{Graph,TestNode,DEFAULT_DOT_OPTIONS};

let mut graph: Graph<TestNode> = Graph::new(3);
graph.add_edge(0, 1).unwrap();
graph.add_edge(0, 2).unwrap();
graph.add_edge(1, 2).unwrap();

// create string of dotfile
let s = graph.to_dot_with_labels(
   DEFAULT_DOT_OPTIONS,
   |_contained, index| format!("Hey {}!", index)
);

// write to file
let mut f = File::create("example.dot").expect("Unable to create file");
f.write_all(s.as_bytes()).expect("Unable to write data");

In this example, example.dot now contains:

graph G{
    bgcolor="transparent";
    fontsize=50;
    node [shape=ellipse, penwidth=1, fontname="Courier", pin=true ];
    splines=true;
    0 1 2 ;
    "0" [label="Hey 0!"];
    "1" [label="Hey 1!"];
    "2" [label="Hey 2!"];
    0 -- 1
    0 -- 2
    1 -- 2
}

Then you can use, e.g.,

foo@bar:~$ circo example.dot -Tpdf > example.pdf

to create a pdf representation from it. Search for graphviz to learn more.

pub fn diameter(&self) -> Option<u32>

returns None if graph not connected or does not contain any vertices

uses repeated breadth first search

pub fn longest_shortest_path_from_index(&self, index: u32) -> Option<u32>

calculate the size of the longest shortest path starting from vertex with index index using breadth first search

pub fn vertex_biconnected_components(
self,
alternative_definition: bool
) -> Vec<usize>

calculate sizes of all binode connected components

returns (reverse) ordered vector of sizes i.e. the biggest component is of size result[0] and the smallest is of size result[result.len() - 1]

destroys the underlying topology and therefore moves self

if you still need your graph, use self.clone().vertex_biconnected_components(false/true) for your calculations

Definition: vertex_biconnected_components(false)

Here, the (vertex) biconnected component of a graph is defined as maximal subset of nodes, where any one node could be removed and the remaining nodes would still be a connected component.

Note

Two vertices connected by an edge are considered to be biconnected, since after the removal of one vertex (and the corresponding edge), only one vertex remains. This vertex is in a connected component with itself.

Alternative Definition: vertex_biconnected_components(true)

If you want to use the alternative definition:

The biconnected component is defined as maximal subset of vertices, where each vertex can be reached by at least two node independent paths

The alternative definition just removes all 2s from the result vector.

Citations

I used the algorithm described in this paper:

J. Hobcroft and R. Tarjan, "Algorithm 447: Efficient Algorithms for Graph Manipulation" Commun. ACM, 16:372-378, 1973, DOI: 10.1145/362248.362272

You can also take a look at:

M. E. J. Newman, "Networks: an Introduction" Oxfort University Press, 2010, ISBN: 978-0-19-920665-0.

pub fn vertex_load(&self, include_endpoints: bool) -> Vec<f64>

Closely related (most of the time equal) to betweeness

calculates vertex_load of all vertices in O(edges * vertices)

calculates the vertex_load for every vertex

defined as how many shortest paths pass through each vertex

variant

vertex_load(true)	includes endpoints in calculation (for a complete graph with N vertices, every node will have vertex_load N - 1)

vertex_load(false)	excludes endpoints in calculation (for a complete graph with N vertices, every node will have vertex_load 0)

Citations

I used the algorithm described in

M. E. J. Newman, "Scientific collaboration networks. II. Shortest paths, weighted networks, and centrality", Phys. Rev. E 64, 016132, 2001, DOI: 10.1103/PhysRevE.64.016132

see also:

M. E. J. Newman, "Erratum: Scientific collaboration networks. II. Shortest paths, weighted networks, and centrality", Phys. Rev. E 73, 039906, 2006, DOI: 10.1103/PhysRevE.73.039906

pub fn transitivity(&self) -> f64

Calculates transitivity of graph

related to cluster coefficient (Note: transitivity and cluster coefficient are similar, but not necessarily equal)

returns NaN, if there are no paths of length two in the graph

Definition

transitivity = (number of closed paths of length two) / (number of paths of length two)

Citations

For the definition see for example:

M. E. J. Newman, "Networks: an Introduction" Oxfort University Press, 2010, ISBN: 978-0-19-920665-0.