Trait net_ensembles::graph_traits::AdjContainer

pub trait AdjContainer<T: Node> where
    Self: Display, {
    fn new(id: u32, node: T) -> Self;

    fn parse_str(to_parse: &str) -> Option<(&str, Self)>
    where
        Self: Sized;

    fn contained<'a>(&'a self) -> &'a T;

    fn contained_mut(&mut self) -> &mut T;

    fn neighbors(&self) -> Iter<u32>;

    fn neighbor_count(&self) -> usize;

    fn id(&self) -> u32;

    fn get_adj_first(&self) -> Option<&u32>;

    fn is_adjacent(&self, other_id: &u32) -> bool;

    fn sort_adj(&mut self);

    unsafe fn clear_edges(&mut self);

    unsafe fn push(&mut self, other: &mut Self) -> Result<(), GraphErrors>;

    unsafe fn remove(&mut self, other: &mut Self) -> Result<(), GraphErrors>;
}

Required methods

fn new(id: u32, node: T) -> Self

fn parse_str(to_parse: &str) -> Option<(&str, Self)> where
Self: Sized,

parse from str

tries to parse a AdjContainer from a str.

returns None if failed

Returns Option((a, b))

a) string slice beginning directly after the part, that was used to parse

b) the AdjContainer resulting form the parsing

fn contained<'a>(&'a self) -> &'a T

return reference to what the AdjContainer contains

fn contained_mut(&mut self) -> &mut T

return mut reference to what the AdjContainer contains

fn neighbors(&self) -> Iter<u32>

returns iterator over indices of neighbors

fn neighbor_count(&self) -> usize

count number of neighbors, i.e. number of edges incident to self

fn id(&self) -> u32

returns id of container

fn get_adj_first(&self) -> Option<&u32>

returns Some(first element from the adjecency List) or None

fn is_adjacent(&self, other_id: &u32) -> bool

check if vertex with other_id is adjacent to self

Note:

(in Graph<T>: id equals the index corresponding to self)

fn sort_adj(&mut self)

Sorting adjecency lists

unsafe fn clear_edges(&mut self)

Remove all edges

Important

will not clear edges of other AdjContainer

only call this if you know exactly what you are doing

unsafe fn push(&mut self, other: &mut Self) -> Result<(), GraphErrors>

What does it do?

Creates edge in self and others adjecency Lists

Why is it unsafe?

No logic to see, if AdjContainer are part of the same graph

Only intended for internal usage

What should I do?

use members of net_ensembles::GenericGraph instead, that handles the logic

unsafe fn remove(&mut self, other: &mut Self) -> Result<(), GraphErrors>

What does it do?

Removes edge in self and others adjecency Lists

Why is it unsafe?

No logic to see, if AdjContainer are part of the same graph

Only intended for internal usage

What should I do?

use members of net_ensembles::GenericGraph instead, that handles the logic

Loading content...