Module net_ensembles::iter

Contains definitions of a few iterators. Not All of them though.

They are returned by a few of the methods, you do not need to build them yourself.

Structs

ContainedIter

Iterator over additional data stored at each vertex in order of indices

ContainedIterMut

same as ContainedIter but mutable

Iterator over mutable additional information stored at vertices

INContainedIterMut

same as NContainedIter but mutable

Iterator over mutable additional information stored at vertices that are directly connected to specific vertex

NContainedIter

Iterator over additional information stored at vertices that are directly connected to specific vertex

NContainedIterMut

same as NContainedIter but mutable

Iterator over mutable additional information stored at vertices that are directly connected to specific vertex

NContainerIter

Iterator over each vertex directly connected with start vertex in adjecency list of vertex index

NIContainedIter

Iterator over additional information + indices stored at vertices that are directly connected to specific vertex

Enums

IterWrapper

Wrapper for iterators