Module net_ensembles::sampling::entropic_sampling

For using entropic sampling after a wang landau simulation

You can store related measurable quantities for looking at correlations etc.

Structs

EntropicSampling
Entropic sampling made easy

EntropicSamplingAdaptive
Entropic sampling made easy

Enums

EntropicErrors
Error states, that entropic sampling, or the creation of EntropicSamplingAdaptive could encounter

Traits

Entropic
Traits for quantities that all Entropic simulations have

EntropicEEH
Helper trait, so that you have to type less

EntropicEnergy
trait to request the current energy from a Entropic simulation

EntropicEnsemble
trait to request a reference to the current (state of the) ensemble

EntropicHist
trait to request the current histogram from a Entropic simulation