Trait net_ensembles::sampling::entropic_sampling::Entropic

pub trait Entropic {
    pub fn step_counter(&self) -> usize;

    pub fn total_steps_accepted(&self) -> usize;

    pub fn total_steps_rejected(&self) -> usize;

    pub fn step_goal(&self) -> usize;

    pub fn log_density(&self) -> Vec<f64, Global>;

    pub fn write_log<W>(&self, writer: W) -> Result<(), Error>
    where
        W: Write;

    pub fn steps_total(&self) -> usize { ... }

    pub fn fraction_accepted_total(&self) -> f64 { ... }

    pub fn fraction_rejected_total(&self) -> f64 { ... }

    pub fn is_finished(&self) -> bool { ... }

    pub fn log_density_base10(&self) -> Vec<f64, Global> { ... }

    pub fn log_density_base(&self, base: f64) -> Vec<f64, Global> { ... }
}

Traits for quantities that all Entropic simulations have

Required methods

pub fn step_counter(&self) -> usize

Counter

how many Entropic steps were performed until now?

pub fn total_steps_accepted(&self) -> usize

How many steps were accepted until now?

this includes steps, that were perfored to find the estimate, i.e., during wang landau

pub fn total_steps_rejected(&self) -> usize

How many steps were rejected until now?

this includes steps, that were perfored to find the estimate, i.e., during wang landau

pub fn step_goal(&self) -> usize

What is the goal to reach?

pub fn log_density(&self) -> Vec<f64, Global>

Current (non normalized) estimate of ln(P(E))

i.e., of the natural logarithm of the probability density function for the requested interval

this is what we are doing the simulations for

pub fn write_log<W>(&self, writer: W) -> Result<(), Error> where
W: Write,

Writes Information about the simulation to a file. E.g. How many steps were performed.

Loading content...
Provided methods

pub fn steps_total(&self) -> usize

Counter

how many markov steps were performed until now?

this includes steps, that were perfored to find the estimate, i.e., during wang landau

pub fn fraction_accepted_total(&self) -> f64

Calculate, which fraction of steps were accepted

this includes steps, that were perfored to find the estimate, i.e., during wang landau

if no steps were performed, it returns f64::NAN

pub fn fraction_rejected_total(&self) -> f64

Calculate, which fraction of steps were rejected

this includes steps, that were perfored to find the estimate, i.e., during wang landau

if no steps were performed, it returns f64::NAN

pub fn is_finished(&self) -> bool

Checks wang landau threshold

log_f <= log_f_threshold

pub fn log_density_base10(&self) -> Vec<f64, Global>

Current (non normalized) estimate of log10(P(E))

i.e., of logarithm with base 10 of the probability density function for the requested interval

this is what we are doing the simulations for

pub fn log_density_base(&self, base: f64) -> Vec<f64, Global>

Current (non normalized) estimate of log_base(P(E))

i.e., of logarithm with arbitrary base of the probability density function for the requested interval

this is what we are doing the simulations for

Loading content...