Enum net_ensembles::sampling::GlueErrors

pub enum GlueErrors {
    BorderCreation(HistErrors),
    EmptyList,
    BinarySearch,
    OutOfBounds,
    NoOverlap,
}

Possible errors that can occur during gluing together WangLandau intervals or Entropic Sampling intervals

Variants

BorderCreation(HistErrors)

original_hist.borders_clone() failed

EmptyList

Nothing to be glued, glue interval list was empty

BinarySearch

Binary search failed - PartialOrd::partial_cmp returned None

OutOfBounds

Glue interval and intervals to be glued do not match

Likely original_hist is to small

NoOverlap

The intervals need to overlap, otherwise no gluing can occur