Enum net_ensembles::sampling::RewlBuilderErr

pub enum RewlBuilderErr {
    NonNormalThreshold,
    Negative,
    Empty,
    HistBinCount,
    SeedError(Error),
    LenMissmatch,
}

Errors

that can arise during the construction of RewlBuilder

Variants

NonNormalThreshold

The threshold for log_f needs to be a normal number.

That basically means: the number is neither zero, infinite, subnormal, or NaN. For more info, see the Documentation

Negative

log_f_threshold must not be negative

Empty

Histogram vector needs to contain at least one entry.

HistBinCount

Each histogram needs to have at least two bins. Though more than two bins are strongly recommended

SeedError(Error)

Unable to seed random number generator

LenMissmatch

Length of histogram vector and ensemble vector has to be the same!