Struct net_ensembles::sampling::examples::coin_flips::CoinFlipSequence

pub struct CoinFlipSequence<R> { /* private fields */ }

A sequence of Coin flips. Contains random Number generator

Implementations

impl<R> CoinFlipSequence<R>where
R: Rng,

pub fn new(n: usize, rng: R) -> CoinFlipSequence<R>

Create new coin flip sequence

length n

use rng as random number generator

impl<R> CoinFlipSequence<R>

pub fn head_count(&self) -> usize

Count how often Head occurs in the Coin flip sequence

pub fn update_head_count(&self, step: &CoinFlipMove, head_count: &mut usize)

Calculate the head count, if a previouse head count of the ensemble and the markov steps leading to the current state are known

head_count is updated

might panic if step was not the markov step leading from the ensemble with head_count to the current ensemble - if it does not panic, the result will be wrong

pub fn max_heads_in_a_row(&self) -> usize

Count many times Head occured in a row

uses maximum value, i.e., for the sequence HHTHHHT it will return 3