Function net_ensembles::sampling::bootstrap_copyable

pub fn bootstrap_copyable<F, R, N1>(
    rng: R,
    samples: usize,
    data: &[N1],
    reduction: F
) -> (f64, f64)where
    F: Fn(&mut [N1]) -> f64,
    R: Rng,
    N1: Copy,

Similar to bootstrap but for stuff that implements Copy. Likely more effient in these cases returns reduced value + estimated error (as sqrt of variance)