Function net_ensembles::sampling::derivative_merged_and_aligned

pub fn derivative_merged_and_aligned<H, Hist>(
    log_prob: Vec<Vec<f64, Global>, Global>,
    hists: Vec<H, Global>,
    log_base: LogBase
) -> Result<ReplicaGlued<Hist>, HistErrors>where
    Hist: HistogramCombine + Histogram,
    H: Borrow<Hist>,

Calculate the probability density function from overlapping intervals

log_prob is a vector of the logarithmic non-normalized probability densities

hists is a vector of the corresponding histograms

LogBase: Which base do the logarithmic probabilities have?

This uses a derviative merge, that works similar to: derivative_merged_log_prob_and_aligned

The ReplicaGlued allows you to easily write the probability density function to a file