Module net_ensembles::sampling::glue

Glue together overlapping intervals of either entropic sampling or wang landau

Use this, if it makes sense to split the range, you are interested in, into multiple intervals

Structs

GlueResult
Glued together probability

Enums

GlueErrors
Possible errors that can occur during gluing together WangLandau intervals or Entropic Sampling intervals

Functions

glue_entropic
Combine multiple WangLandau intervals to get the probability distribution of the whole interval

glue_wl
Combine multiple WangLandau intervals to get the probability distribution of the whole interval

norm_log10_sum_to_1
Normalize log10 probability density