Function net_ensembles::sampling::glue::glue_entropic

pub fn glue_entropic<Entr, HIST, T>(
    list: &[Entr], 
    original_hist: &HIST
) -> Result<GlueResult<T>, GlueErrors> where
    T: PartialOrd<T>,
    HIST: Histogram + HistogramVal<T>,
    Entr: EntropicHist<HIST>, 

Combine multiple WangLandau intervals to get the probability distribution of the whole interval

list: Vector/slice of Wang landau distributions.

Restrictions

original_hist has to contain all the borders of the histograms. Meaning: The size of a bin has to be constant among all histograms of the list. If it is not, you might get an error, or you might get wrong results. I do not check for this exaustingly.

There is an easy way to make sure, that you dont get problems here: Create the original_histfirst. Then create the other Histograms using theHistogramPartion` trait. This is the intended way. But as long as the borders and bin sizes match, this function will work properly

Understanding returned Parameters (OK(..)):