Trait net_ensembles::sampling::histogram::HistogramIntervalDistance

pub trait HistogramIntervalDistance<T> {
    pub fn interval_distance_overlap<V>(
        &self, 
        val: V, 
        overlap: NonZeroUsize
    ) -> usize
    where
        V: Borrow<T>;
}

Distance metric for how far a value is from a valid interval

Required methods

pub fn interval_distance_overlap<V>(
&self,
val: V,
overlap: NonZeroUsize
) -> usize where
V: Borrow<T>,

Distance metric for how far a value is from a valid interval

partitions in more intervals, checks which bin interval a bin corresponds to and returns distance of said interval to the target interval

used for heuristiks

overlap should be bigger 0, otherwise it will be set to 1

Loading content...