Trait net_ensembles::sampling::histogram::HistogramVal

pub trait HistogramVal<T>: Histogram {
    pub fn get_bin_index<V>(&self, val: V) -> Result<usize, HistErrors>
    where
        V: Borrow<T>;

    pub fn borders_clone(&self) -> Result<Vec<T, Global>, HistErrors>;

    pub fn is_inside<V>(&self, val: V) -> bool
    where
        V: Borrow<T>;

    pub fn not_inside<V>(&self, val: V) -> bool
    where
        V: Borrow<T>;

    pub fn first_border(&self) -> T;

    pub fn second_last_border(&self) -> T;

    pub fn distance<V>(&self, val: V) -> f64
    where
        V: Borrow<T>;

    pub fn count_val<V>(&mut self, val: V) -> Result<usize, HistErrors>
    where
        V: Borrow<T>,
    { ... }
}

trait used for mapping values of arbitrary type T to bins

used to create a histogram

Required methods

pub fn get_bin_index<V>(&self, val: V) -> Result<usize, HistErrors> where
V: Borrow<T>,

convert val to the respective histogram index

pub fn borders_clone(&self) -> Result<Vec<T, Global>, HistErrors>

binning borders

the borders used to bin the values

any val which fullfills self.border[i] <= val < self.border[i + 1] will get index i.

Note that the last border is exclusive

pub fn is_inside<V>(&self, val: V) -> bool where
V: Borrow<T>,

does a value correspond to a valid bin?

pub fn not_inside<V>(&self, val: V) -> bool where
V: Borrow<T>,

opposite of is_inside

pub fn first_border(&self) -> T

get the left most border (inclusive)

pub fn second_last_border(&self) -> T

get second last border from the right

should be the same as let b = self.borders_clone().expect("overflow"); assert_eq!(self.second_last_border(), b[b.len()-2])

pub fn distance<V>(&self, val: V) -> f64 where
V: Borrow<T>,

calculates some sort of absolute distance to the nearest valid bin

any invalid numbers (like NAN or INFINITY) should have the highest distance possible

if a value corresponds to a valid bin, the distance should be zero

Loading content...
Provided methods

pub fn count_val<V>(&mut self, val: V) -> Result<usize, HistErrors> where
V: Borrow<T>,

count val. Some(index), if inside of hist, None if val is invalid

Loading content...