Function net_ensembles::sampling::rewl::merged_log_prob

pub fn merged_log_prob<Ensemble, R, Hist, Energy, S, Res>(
    rewls: &[ReplicaExchangeWangLandau<Ensemble, R, Hist, Energy, S, Res>]
) -> Result<(Vec<f64, Global>, Hist), HistErrors> where
    Hist: HistogramVal<Energy> + HistogramCombine + Send + Sync,
    Energy: PartialOrd<Energy>, 

Will calculate the merged log (base e) probability density. Also returns the corresponding histogram.

If an interval has multiple walkers, their probability will be merged before all probabilities are aligned

rewls does not need to be sorted in any way

Errors

will return HistErrors::EmptySlice if the rees slice is empty

will return other HistErrors if the intervals have no overlap