Struct net_ensembles::sampling::HeatmapF64Mean

pub struct HeatmapF64Mean<HistX, HistY> { /* private fields */ }

Heatmap with mean of y-axis

stores heatmap in row-major order: the rows of the heatmap are contiguous, and the columns are strided

enables you to quickly create a heatmap

you can create gnuplot scripts to plot the heatmap

for each x-axis bin, the y-axis mean is calculated

…

Difference to HeatmapF64

HeatmapF64 does not contain the averages for th y-axis, but can be transposed and also used for Y-Histograms which take types which do not implement AsPrimitive

Implementations

impl<HistX, HistY> HeatmapF64Mean<HistX, HistY>

pub fn heatmap(&self) -> &HeatmapF64<HistX, HistY>

Internal HeatmapF64

impl<HistX, HistY> HeatmapF64Mean<HistX, HistY>where
HistX: Histogram,
HistY: Histogram,

pub fn new(hist_x: HistX, hist_y: HistY) -> HeatmapF64Mean<HistX, HistY>

Create a heatmap

creates new instance

hist_x defines the bins along the x-axis

hist_y defines the bins along the y-axis

pub fn count_inside_heatmap<X, Y, A, B>(
&mut self,
x_val: A,
y_val: B,
weight: f64
) -> Result<(usize, usize), HeatmapError>where
HistX: HistogramVal<X>,
HistY: HistogramVal<Y>,
A: Borrow<X>,
B: Borrow<Y>,
Y: AsPrimitive<f64>,

Update Heatmap

similar to count of HeatmapF64

This time, however, any value that is out of bounds will be ignored for the calculation of the mean of the y-axis, meaning also values which correspond to a valid x-bin will be ignored, if their y-value is not inside the Y Histogram. The mean respects the weight

pub fn count<X, Y, A, B>(
&mut self,
x_val: A,
y_val: B,
weight: f64
) -> Result<(usize, usize), HeatmapError>where
HistX: HistogramVal<X>,
HistY: HistogramVal<Y>,
A: Borrow<X>,
B: Borrow<Y>,
Y: AsPrimitive<f64>,

Update heatmap

Corresponds to count of HeatmapU

The difference is, that the mean of the y-axis is updated as long as y_val is finite and x_val is in bounds (because the mean is calculated for each bin in the x direction separately). The calculated average respects the weight

pub fn mean_slice(&self) -> &[WeightedMean]

Internal slice for mean

The mean is calculated from this slice

The mean corresponds to the bins of the x-axis

pub fn mean_iter(&self) -> impl Iterator<Item = f64>

Iterate over the calculated mean

iterates over the means

The mean corresponds to the bins of the x-axis

if a bin on the x-axis has no entries, the corresponding mean will be f64::NAN

pub fn mean(&self) -> Vec<f64, Global>

Get a mean vector

The entries are the means corresponds to the bins of the x-axis

if a bin on the x-axis has no entries, the corresponding mean will be f64::NAN

Note

If you want to iterate over the mean values, use mean_iter instead

pub fn gnuplot_quick<W>(&self, writer: W) -> Result<(), Error>where
W: Write,

Create a gnuplot script to plot your heatmap

writer: The gnuplot script will be written to this

gnuplot_output_name: how shall the file, created by executing gnuplot, be called? Ending of file will be set automatically

Note

This is the same as calling gnuplot with default GnuplotSettings and default GnuplotPointSettings

The default axis are the bin indices, which, e.g, means they always begin at 0. You have to set the axis via the GnuplotSettings

pub fn gnuplot<W, P, GS>(
&self,
writer: W,
settings: GS,
points: P
) -> Result<(), Error>where
W: Write,
P: Borrow<GnuplotPointSettings>,
GS: Borrow<GnuplotSettings>,

Create a gnuplot script to plot your heatmap

This function writes a file, that can be plotted in the terminal via gnuplot

gnuplot gnuplot_file

Parameter:

writer: writer gnuplot script will be written to

gnuplot_output_name: how shall the file, created by executing gnuplot, be called? File suffix (ending) will be set automatically

settings: Here you can set the axis, choose between terminals and more. I recommend that you take a look at GnuplotSettings

point_color: the mean (in y-direction) will be plotted as points in the heatmap. Here you can choose the point color

Notes

The default axis are the bin indices, which, e.g, means they always begin at 0. You have to set the axis via the GnuplotSettings