Struct net_ensembles::sampling::HistogramFloat

pub struct HistogramFloat<T> { /* fields omitted */ }

Generic Histogram struct

Implementations

impl<T> HistogramFloat<T>

pub fn borders(&self) -> &Vec<T, Global>

similar to self.borders_clone but does not allocate memory

impl<T> HistogramFloat<T> where
T: Float + PartialOrd<T> + FromPrimitive,

pub fn new(
left: T,
right: T,
bins: usize
) -> Result<HistogramFloat<T>, HistErrors>

Create a new Historgram

right exclusive, left inclusive

if you want right to behave (almost) the same as an inclusive border, consider using new(left, right + T::EPSILON, bins) (make sure, that adding Epsilon actually changes the value!)

pub fn interval_length(&self) -> T

Returns the length of the interval