Trait net_ensembles::sampling::EntropicEnergy

pub trait EntropicEnergy<Energy>: Entropic {
    pub fn energy(&self) -> &Energy;
}

trait to request the current energy from a Entropic simulation

None if the energy was not calculated yet

See also EntropicEEH

Required methods

pub fn energy(&self) -> &Energy

returns the last accepted Energy calculated None if no energy was calculated yet

Loading content...