Trait net_ensembles::sampling::HistogramCombine

pub trait HistogramCombine: Sized {
    fn encapsulating_hist<S>(hists: &[S]) -> Result<Self, HistErrors>
    where
        S: Borrow<Self>;
    fn align<S>(&self, right: S) -> Result<usize, HistErrors>
    where
        S: Borrow<Self>;
}

Used to get a histogram, which contains the smaller histograms

Required Methods

fn encapsulating_hist<S>(hists: &[S]) -> Result<Self, HistErrors>where
S: Borrow<Self>,

Create a histogram, which encapsulates the histograms passed

possible errors

bin size of histograms is unequal

bins do not align

fn align<S>(&self, right: S) -> Result<usize, HistErrors>where
S: Borrow<Self>,

Get bin difference between histograms

index of bin of self corresponding to the leftest bin of right