Trait net_ensembles::sampling::HistogramIntervalDistance

pub trait HistogramIntervalDistance<T> {
    fn interval_distance_overlap<V>(&self, val: V, overlap: NonZeroUsize) -> usize
    where
        V: Borrow<T>;
}

Distance metric for how far a value is from a valid interval

Required Methods

fn interval_distance_overlap<V>(&self, val: V, overlap: NonZeroUsize) -> usizewhere
V: Borrow<T>,

Distance metric for how far a value is from a valid interval

partitions in more intervals, checks which bin interval a bin corresponds to and returns distance of said interval to the target interval

used for heuristiks

overlap should be bigger 0, otherwise it will be set to 1