Trait net_ensembles::sampling::MarkovChain

pub trait MarkovChain<S, Res> {
    pub fn undo_step(&mut self, step: &S) -> Res;

    pub fn undo_step_quiet(&mut self, step: &S);

    pub fn m_step(&mut self) -> S;

    pub fn m_steps(&mut self, count: usize, steps: &mut Vec<S, Global>) { ... }

    pub fn m_steps_quiet(&mut self, count: usize) { ... }

    pub fn m_step_acc<Acc, AccFn>(&mut self, acc: &mut Acc, acc_fn: AccFn) -> S
    where
        AccFn: FnMut(&Self, &S, &mut Acc),
    { ... }

    pub fn m_steps_acc<Acc, AccFn>(
        &mut self, 
        count: usize, 
        steps: &mut Vec<S, Global>, 
        acc: &mut Acc, 
        acc_fn: AccFn
    )
    where
        AccFn: FnMut(&Self, &S, &mut Acc),
    { ... }

    pub fn m_steps_acc_quiet<Acc, AccFn>(
        &mut self, 
        count: usize, 
        acc: &mut Acc, 
        acc_fn: AccFn
    )
    where
        AccFn: FnMut(&Self, &S, &mut Acc),
    { ... }

    pub fn undo_steps(&mut self, steps: &[S], res: &mut Vec<Res, Global>) { ... }

    pub fn undo_steps_quiet(&mut self, steps: &[S]) { ... }
}

Create a markov chain by doing markov steps

Required methods

pub fn undo_step(&mut self, step: &S) -> Res

undo a markov step, return result-state

if you want to undo more than one step see undo_steps

pub fn undo_step_quiet(&mut self, step: &S)

undo a markov, panic on invalid result state

for undoing multiple steps see undo_steps_quiet

pub fn m_step(&mut self) -> S

Markov step

use this to perform a markov step step

for doing multiple markov steps at once, use m_steps

Loading content...
Provided methods

pub fn m_steps(&mut self, count: usize, steps: &mut Vec<S, Global>)

Markov steps

use this to perform multiple markov steps at once

steps can be used to undo the steps with self.undo_steps(steps)

steps will be emptied before step is performed

pub fn m_steps_quiet(&mut self, count: usize)

Markov steps without return

use this to perform multiple markov steps at once

only use this if you know that you do not want to undo the steps

you cannot undo this steps, but therefore it does not need to allocate a vector for undoing steps

pub fn m_step_acc<Acc, AccFn>(&mut self, acc: &mut Acc, acc_fn: AccFn) -> S where
AccFn: FnMut(&Self, &S, &mut Acc),

Accumulating markov step

this calculates something while performing the markov chain, e.g., the current energy, which can be more efficient then calculating it from scratch afterwards

pub fn m_steps_acc<Acc, AccFn>(
&mut self,
count: usize,
steps: &mut Vec<S, Global>,
acc: &mut Acc,
acc_fn: AccFn
) where
AccFn: FnMut(&Self, &S, &mut Acc),

Accumulating markov steps

this calculates something while performing the markov chain, e.g., the current energy which can be more efficient then calculating it from scratch afterwards

pub fn m_steps_acc_quiet<Acc, AccFn>(
&mut self,
count: usize,
acc: &mut Acc,
acc_fn: AccFn
) where
AccFn: FnMut(&Self, &S, &mut Acc),

Accumulating markov steps

this calculates something while performing the markov chain, e.g., the current energy which can be more efficient then calculating it from scratch afterwards

quiet step, i.e., you will not be able to undo the step

pub fn undo_steps(&mut self, steps: &[S], res: &mut Vec<Res, Global>)

Undo markov steps

Note: uses undo_step in correct order and returns result

Important:

look at specific implementation of undo_step, every thing mentioned there applies to each step

pub fn undo_steps_quiet(&mut self, steps: &[S])

Undo markov steps

Note: uses undo_step_quiet in correct order

Important:

look at specific implementation of undo_step_quiet, every thing mentioned there applies to each step

Loading content...