Trait net_ensembles::sampling::WangLandauEnsemble

pub trait WangLandauEnsemble<E>: WangLandau {
    fn ensemble(&self) -> &E;
    unsafe fn ensemble_mut(&mut self) -> &mut E;
}

trait to request a reference to the current (state of the) ensemble

See also WangLandauEEH

Required Methods

fn ensemble(&self) -> &E

return reference to current state of ensemble

unsafe fn ensemble_mut(&mut self) -> &mut E

mutable reference to current state

unsafe you should not make any changes, that effect the energy state of the system. otherwise the WangLandau simulations will give false results!