Trait net_ensembles::sampling::wang_landau::WangLandauHist

pub trait WangLandauHist<Hist>: WangLandau {
    fn hist(&self) -> &Hist;
}

trait to request the current histogram from a WangLandau simulation

Note: The histogram will likely be reset multiple times during a simulation

See also WangLandauEEH

Required Methods

fn hist(&self) -> &Hist

returns current histogram

Note: histogram will be reset multiple times during the simulation

please refere to the papers