Struct net_ensembles::spacial::SpacialEnsemble

pub struct SpacialEnsemble<T, R> where
    T: Node,  { /* fields omitted */ }

Implements a special Ensemble

You can generate a dot file which includes special information.

NOTE You should use neato for that to work

Implementations

impl<T, R> SpacialEnsemble<T, R> where
T: Node,
R: Rng,

pub fn new(n: usize, rng: R, f: f64, alpha: f64) -> Self

Generate a new Spacial ensemble with

n nodes

rng as random number generator

f - see paper

alpha - see paper