Type Definition net_ensembles::spacial::SpacialGraph

pub type SpacialGraph<T> = GenericGraph<T, SpacialNodeContainer<T>>;

Type definiton for convinience. This is used to implement the spacial ensemble

Implementations

impl<T> SpacialGraph<T>

pub fn distance(&self, i: usize, j: usize) -> Option<f64>where
SpacialNodeContainer<T>: AdjContainer<T>,

Euclidean distance between two vertices

Calculates the distance between the vertices corresponding to the indices i and j

None if any of the indices is out of bounds