Struct net_ensembles::BAensemble

pub struct BAensemble<T, R> where
    T: Node,
    R: Rng,  { /* fields omitted */ }

Implements a Barabási-Albert Graph ensemble

Implementations

impl<T, R> BAensemble<T, R> where
T: Node + SerdeStateConform,
R: Rng,

pub fn new(n: usize, rng: R, m: usize, source_n: usize) -> Self

Initialize

create simplest form of Barabási-Albert graph

m = 1

n: Number of nodes, n > 1 has to be true panics otherwise

rng: Rng to use

pub fn new_from_graph<B>(n: usize, rng: R, m: usize, source_graph: B) -> Self where
B: Borrow<Graph<T>>,

Generate a new BA graph ensemble with a specified source graph

panics if source_graph contains any vertices with degree 0

m: how many edges should each newly added vertex have originally

rng: Random number generator

n: Number of nodes, n > source_graph.vertex_count() has to be true panics otherwise

pub fn new_from_generic_graph<A2, B>(
n: usize,
rng: R,
m: usize,
generic_source_graph: B
) -> Self where
A2: AdjContainer<T>,
B: Borrow<GenericGraph<T, A2>>,

Generate a new BA graph ensemble with a specified generic source graph

panics if generic_source_graph contains any vertices with degree 0

m: how many edges should each newly added vertex have originally

rng: Random number generator

n: Number of nodes, n > source_graph.vertex_count() has to be true panics otherwise

pub fn source_graph(&self) -> &Graph<T>

get reference to original graph, which is at the core of the BA graph