Struct net_ensembles::ConfigurationModel

pub struct ConfigurationModel<T, R> where
    T: Node,  { /* fields omitted */ }

Generate networks with a given degree distribution

the degree of each vertex is fixed (see self.degree_vec), while the actual edges will be drawn randomly

No self loops allowed

Implementations

impl<T, R> ConfigurationModel<T, R> where
T: Node,

pub fn degree_vec(&self) -> &Vec<usize>

Get reference to the degree vector of the vertices, faster than self.graph().degree_vec(), since the former has to construct the vector, while the latter just returns a reference to an existing vector

impl<T, R> ConfigurationModel<T, R> where
T: Node,
R: Rng,

pub fn from_const(constant: usize, size: usize, rng: R) -> Option<Self>

create configuration model from a constant degree

drawn graphs will consist of degree_vec.len() vertices, where a vertex i will have degree degree_vec[i]

size: number of nodes in the resulting graphs

returns None if resulting degree vector is invalid

pub fn from_generic_graph<T1, A1>(
generic_graph: &GenericGraph<T1, A1>,
rng: R
) -> Self where
T1: Node,
A1: AdjContainer<T1>,

create ConfigurationModel from a generic graph

same as from_vec_unchecked, but creates degree vector from a generic graph

pub fn clone_from_generic_graph<A>(
generic_graph: &GenericGraph<T, A>,
rng: R
) -> Self where
T: Clone,
A: AdjContainer<T>,

create ConfigurationModel from a generic graph and clones underlying Data

degree_vec will be extracted from generic_graph

Data will be cloned from the generic_graph

returns ConfigurationModel model

model.graph() will have the same topology as generic_graph after this creation. This will of cause change, if you call randomize or do markov steps

pub fn from_vec(degree_vec: Vec<usize>, rng: R) -> Option<Self>

create configuration model from a degree vector

drawn graphs will consist of degree_vec.len() vertices, where a vertex i will have degree degree_vec[i]

returns None if degree vector is invalid

pub fn from_vec_unchecked(degree_vec: Vec<usize>, rng: R) -> Self

create configuration model from a degree vector

same as from_vec, but it does not check if the degree_vec is valid - that is on you now

pub fn degree_vec_is_valid(degree_vec: &Vec<usize>) -> bool

check if a vector is a vaild degree distribution

sum of degree_vec needs to be even

degree_vec.len() has to be greater than 1

no entry can request a degree larger than degree_vec.len()-2

pub fn assert_degree_vec_valid(degree_vec: &Vec<usize>)

asserts, that a vector is a vaild degree distribution

similar to degree_vec_is_valid, but asserts instead

intended for debugging: see why the degree_vec is invalid

pub fn swap_degree_vec(&mut self, new_degree_vec: Vec<usize>) -> Vec<usize>

Swaps the degree vector for a new one and draws a new network accordingly

Note new_degree_vec.len() has to be of the same length as self.degree_vec.len() will panic otherwise

panics if new_degree_vec is invalid

returns previous degree_vec

pub fn swap_degree_vec_unchecked(
&mut self,
new_degree_vec: Vec<usize>
) -> Vec<usize>

Swaps the degree_vec for a new one and draws a new network accordingly

same as swap_degree_vec but does not assert, that the degree vector is valid

panics if self.degree_vec.len() != new_degree_vec.len() (the only thing still checked for, as that check is really really cheap)

pub fn degree_vec_from_generic_graph<T1, A1>(
&mut self,
generic_graph: &GenericGraph<T1, A1>
) where
T1: Node,
A1: AdjContainer<T1>,

Use the degree vector of a generic graph

asserts, that generic_graph and self.graph() have the same number of vertices

similar to self.swap_degree_vec_unchecked(generic_graph.degree_vec()) but does not create new vector and as such does not return old degree vector. If you need the old degree vector, you can use self.degree_vec().clone() before calling this method