Module net_ensembles::sw

Small-world ensemble

In this specific small-world ensemble each vertex has at least degree 2. That means, this small-world ensemble will never exhibit leaves.

I implemented the same model, as I used in my paper

Yannick Feld and Alexander K. Hartmann, "Large-deviations of the basin stability of power grids" Chaos 29:113113 (2019), DOI: 10.1063/1.5121415

where it is described in more detail.

You can find a list of my publications on my homepage.

Citations

D. J. Watts and S. H. Strogatz, "Collective dynamics on 'small-world' networks," Nature 393, 440-442 (1998), DOI: 10.1038/30918

Structs

SwEnsemble
Implements small-world graph ensemble

Enums

SwChangeState
Returned by markov steps