Struct net_ensembles::sw_graph::SwEdge

pub struct SwEdge { /* fields omitted */ }

Edge of small world network

Implementations

impl SwEdge

pub fn to(&self) -> &usize

Where does the edge point to, i.e., to which node does it connect?

pub fn is_root(&self) -> bool

Is the edge a root edge?

A root edge is an edge which will allways be connected to the current node. Where it connects to can change, where it connects from cannot.

pub fn is_at_root(&self) -> bool

Is the edge at its root position?

A root edge is an edge which will allways be connected to the current node. Where it connects to can change, where it connects from cannot.

A root edge can be reset (i.e., where it connects to is reset) to its original neighbor (the next, or second next neighbor)

This checks, if the edge still connects to where it would be reset to anyway

pub fn is_long_ranging_root(&self) -> bool

checks root edge it it is long ranging

is it a root edge and if yes, is it a long ranging root edge?