Type Definition net_ensembles::sw_graph::SwGraph

pub type SwGraph<T> = GenericGraph<T, SwContainer<T>>;

specific GenericGraph used for small-world ensemble

Implementations

impl<T> SwGraph<T>where
T: Node + SerdeStateConform,

pub fn reset_edge(&mut self, index0: usize, index1: usize) -> SwChangeState

Reset small-world edge to its root state

panics if index out of bounds

in debug: panics if index0 == index1

pub fn rewire_edge(
&mut self,
index0: usize,
index1: usize,
index2: usize
) -> SwChangeState

Rewire edges

rewire edge (index0, index1) to (index0, index2)

panics

if indices are out of bounds

in debug: panics if index0 == index2

edge (index0, index1) has to be rooted at index0, else will panic in debug mode

pub fn count_nodes_with_long_ranging_edges(&self) -> usize

How many nodes have long ranging edges?

counts how many nodes have long ranging edges

A long ranging edge is defined as an edge, where is_at_root returns true, i.e., which is not in ist original ring configuration

pub fn frac_nodes_wlre(&self) -> f64

Fraction of nodes which have long ranging edges

A long ranging edge is defined as an edge, where is_at_root returns true, i.e., which is not in ist original ring configuration

pub fn count_long_ranging_edges(&self) -> usize

How many long ranging edges are there in the Graph?

A long ranging edge is defined as an edge, where is_at_root returns true, i.e., which is not in ist original ring configuration

pub fn frac_long_ranging_edges(&self) -> f64

Fraction of long ranging edges in the Graph?

A long ranging edge is defined as an edge, where is_at_root returns false, i.e., which is not in ist original ring configuration

this is: #longranging/#total