Module net_ensembles::traits

You should use net_ensembles::traits::*

I recommend doing so for a smooth experience

contains traits you should use for accessing complete functionallity

Re-exports

pub use crate::iter::IterWrapper;

Traits

AdjContainer
Defines methods all adjecency containers should have such that GenericGraph can use it

GraphIterators
Collection of Graph iterators

GraphIteratorsMut
Collection mut Graph iterators

HasRng
Access internal random number generator

MarkovChain
Create a markov chain by doing markov steps

Node
What every node should be able to do

SerdeStateConform

intermediate trait used for trait bounds

if feature "serde_support" is used, there is a blanked implementation for all types that implement serde::Serialize

else there is a blanked implementation for all types

SimpleSample
For easy sampling of your ensemble

WithGraph
unify graph ensembles in a trait