Enum net_ensembles::traits::GraphErrors

pub enum GraphErrors {
    EdgeExists,
    EdgeDoesNotExist,
}

Error messages

Variants

EdgeExists

somehow, the existing of the edge is a problem

Did you try to add an edge, which is already present?

EdgeDoesNotExist

ERROR 404: Edge not found ;)

Did you try to delete a non existing edge?

Implementations

impl GraphErrors

pub fn to_str(&self) -> &'static str

get error message as &str, for printing etc.