Trait net_ensembles::traits::AdjList

pub trait AdjList<Edge> {
    fn edges(&self) -> &[Edge];
}

Get the adjacency list of a AdjContainer

Required Methods

fn edges(&self) -> &[Edge]

Intended for all AdjContainer, which store all their edges in a continouse slice/vector. This will return a reference to the edge slice, which you can use to iterate or do other stuff