Trait net_ensembles::traits::Ensemble

pub trait Ensemble<S, Res> {
    fn undo_step(&mut self, step: S) -> Res;

    fn undo_step_quiet(&mut self, step: S);

    fn randomize(&mut self);

    fn mc_step(&mut self) -> S;

    fn mc_steps(&mut self, count: usize) -> Vec<S> { ... }

    fn simple_sample<F>(&mut self, times: usize, f: F)
    where
        F: FnMut(&Self),
    { ... }

    fn simple_sample_vec<F, G>(&mut self, times: usize, f: F) -> Vec<G>
    where
        F: FnMut(&Self) -> G,
    { ... }

    fn undo_steps(&mut self, steps: Vec<S>) -> Vec<Res> { ... }

    fn undo_steps_quiet(&mut self, steps: Vec<S>) { ... }
}

Sampling an Ensemble

includes:

monte carlo steps

simple sampling

Required methods

fn undo_step(&mut self, step: S) -> Res

undo a monte carlo step, return result-state

if you want to undo more than one step see undo_steps

fn undo_step_quiet(&mut self, step: S)

undo a monte carlo step, panic on invalid result state

for undoing multiple steps see undo_steps_quiet

fn randomize(&mut self)

Randomizes self according to model

this is intended for creation of initial sample

used in simple_sample and simple_sample_vec

fn mc_step(&mut self) -> S

Monte Carlo step

use this to perform a Monte Carlo step

for doing multiple mc steps at once, use mc_steps

Loading content...
Provided methods

fn mc_steps(&mut self, count: usize) -> Vec<S>

Monte Carlo steps

use this to perform multiple Monte Carlo steps at once

result Vec<S> can be used to undo the steps with self.undo_steps(result)

fn simple_sample<F>(&mut self, times: usize, f: F) where
F: FnMut(&Self),

do the following times times:

f(self)

self.randomize()

fn simple_sample_vec<F, G>(&mut self, times: usize, f: F) -> Vec<G> where
F: FnMut(&Self) -> G,

do the following times times:

res = f(self)

self.randomize()

res is collected into Vector

fn undo_steps(&mut self, steps: Vec<S>) -> Vec<Res>

Undo Monte Carlo steps

Note: uses undo_step in correct order and returns result

Important:

look at specific implementation of undo_step, every thing mentioned there applies to each step

fn undo_steps_quiet(&mut self, steps: Vec<S>)

Undo Monte Carlo steps

Note: uses undo_step_quiet in correct order

Important:

look at specific implementation of undo_step_quiet, every thing mentioned there applies to each step

Loading content...