Trait net_ensembles::traits::MarkovChain

pub trait MarkovChain<S, Res> {
    fn undo_step(&mut self, step: S) -> Res;
    fn undo_step_quiet(&mut self, step: S);
    fn m_step(&mut self) -> S;

    fn m_steps(&mut self, count: usize) -> Vec<S> { ... }
    fn undo_steps(&mut self, steps: Vec<S>) -> Vec<Res> { ... }
    fn undo_steps_quiet(&mut self, steps: Vec<S>) { ... }
}

Create a markov chain by doing markov steps

Required Methods

fn undo_step(&mut self, step: S) -> Res

undo a markov step, return result-state

if you want to undo more than one step see undo_steps

fn undo_step_quiet(&mut self, step: S)

undo a markov, panic on invalid result state

for undoing multiple steps see undo_steps_quiet

fn m_step(&mut self) -> S

Markov step

use this to perform a markov step step

for doing multiple markov steps at once, use m_steps

Provided Methods

fn m_steps(&mut self, count: usize) -> Vec<S>

Markov steps

use this to perform multiple markov steps at once

result Vec<S> can be used to undo the steps with self.undo_steps(result)

fn undo_steps(&mut self, steps: Vec<S>) -> Vec<Res>

Undo markov steps

Note: uses undo_step in correct order and returns result

Important:

look at specific implementation of undo_step, every thing mentioned there applies to each step

fn undo_steps_quiet(&mut self, steps: Vec<S>)

Undo markov steps

Note: uses undo_step_quiet in correct order

Important:

look at specific implementation of undo_step_quiet, every thing mentioned there applies to each step