Trait net_ensembles::traits::Node

pub trait Node where
    Self: Clone, {
    fn new_from_index(index: u32) -> Self;

    fn make_string(&self) -> Option<String>;

    fn parse_str(_to_parse: &str) -> Option<(&str, Self)>
    where
        Self: Sized;
}

What every node should be able to do

Required methods

fn new_from_index(index: u32) -> Self

how to construct a blank object

fn make_string(&self) -> Option<String>

Override this, if you want to store the network

fn parse_str(_to_parse: &str) -> Option<(&str, Self)> where
Self: Sized,

Override this, if you want to load the stored network

Loading content...