Struct net_ensembles::er_c::ErEnsembleC

pub struct ErEnsembleC<T: Node, R: Rng> { /* private fields */ }

Implements Erdős-Rényi graph ensemble

variable number of edges

targets a connectivity

Implementations

impl<T: Node, R: Rng> ErEnsembleC<T, R>

pub fn new(n: u32, c_target: f64, rng: R) -> Self

Initialize

create new ErEnsembleC with:

n vertices

target connectivity c_target

rng is consumed and used as random number generator in the following

internally uses Graph<T>::new(n)

generates random edges according to ER model

pub fn target_connectivity(&self) -> f64

returns target connectivity

Explanation

The target connectivity c_target is used to calculate the probability p, that any two vertices i and j (where i != j) are connected.

p = c_target / (N - 1) where N is the number of vertices in the graph

pub fn set_target_connectivity(&mut self, c_target: f64)

set new value for target connectivity

Note

will only set the value (and probability), which will be used from now on

if you also want to create a new sample, call randomize afterwards

pub fn sort_adj(&mut self)

Sort adjecency lists

If you depend on the order of the adjecency lists, you can sort them

Performance

internally uses pattern-defeating quicksort as long as that is the standard

sorts an adjecency list with length d in worst-case: O(d log(d))

is called for each adjecency list, i.e., self.vertex_count() times

pub fn graph(&self) -> &Graph<T>

returns reference to the underlying topology aka, the Graph<T>

Use this to call functions regarding the topology

pub fn at(&self, index: usize) -> &T

access additional information at vertex

pub fn at_mut(&mut self, index: usize) -> &mut T

mutable access of additional information at index