Struct net_ensembles::er_m::ErEnsembleM

pub struct ErEnsembleM<T: Node, R: Rng>where
    T: Node,
    R: Rng,{ /* private fields */ }

Implements Erdős-Rényi graph ensemble

Constant number of edges

Note simple sampling of this ensemble is somewhat inefficient right now - I might change it in the future, though that will change the results of the simple sampling (Not on average of cause)

for simple sampling look at SimpleSample trait

for markov steps look at MarkovChain trait

Other

for topology functions look at GenericGraph

to access underlying topology or manipulate additional data look at WithGraph trait

to use or swap the random number generator, look at HasRng trait

Save and load example

only works if feature "serde_support" is enabled

Note: "serde_support" is enabled by default

I need the #[cfg(feature = "serde_support")] to ensure the example does compile if

you can do not have to use serde_json, look here for more info you opt out of the default feature

use net_ensembles::traits::*; // I recommend always using this
use serde_json;
use rand_pcg::Pcg64;
use net_ensembles::{ErEnsembleM, EmptyNode, rand::SeedableRng};
use std::fs::File;

let rng = Pcg64::seed_from_u64(95);
// create Erdős-Rényi ensemble with 200 vertices and 600 edges
let er_ensemble = ErEnsembleM::<EmptyNode, Pcg64>::new(200, 600, rng);

#[cfg(feature = "serde_support")]
{
    // storing the ensemble in a file:

    let er_m_file = File::create("store_ER_m.dat")
          .expect("Unable to create file");

    // or serde_json::to_writer(er_m_file, &er_ensemble);
    serde_json::to_writer_pretty(er_m_file, &er_ensemble);

    // loading ensemble from file:

    let mut read = File::open("store_ER_m.dat")
        .expect("Unable to open file");

    let er: ErEnsembleM::<EmptyNode, Pcg64> = serde_json::from_reader(read).unwrap();
}

Implementations

impl<T, R> ErEnsembleM<T, R>where
T: Node + SerdeStateConform,
R: Rng,

pub fn new(n: u32, m: usize, rng: R) -> Self

Initialize

create new ErEnsembleM graph with:

n vertices

m edges

rng is consumed and used as random number generator in the following

internally uses Graph<T>::new(n)

generates random edges according to ER model

pub fn get_m(&self) -> usize

Return total number of edges

pub fn sort_adj(&mut self)

Sort adjecency lists

If you depend on the order of the adjecency lists, you can sort them

Performance

internally uses pattern-defeating quicksort as long as that is the standard

sorts an adjecency list with length d in worst-case: O(d log(d))

is called for each adjecency list, i.e., self.vertex_count() times