Struct net_ensembles::er_m::ErEnsembleM

pub struct ErEnsembleM<T: Node, R: Rng> { /* private fields */ }

Implements Erdős-Rényi graph ensemble

Constant number of edges

Implementations

impl<T: Node, R: Rng> ErEnsembleM<T, R>

pub fn new(n: u32, m: usize, rng: R) -> Self

Initialize

create new ErEnsembleM graph with:

n vertices

m edges

rng is consumed and used as random number generator in the following

internally uses Graph<T>::new(n)

generates random edges according to ER model

pub fn get_m(&self) -> usize

Return total number of edges

pub fn sort_adj(&mut self)

Sort adjecency lists

If you depend on the order of the adjecency lists, you can sort them

Performance

internally uses pattern-defeating quicksort as long as that is the standard

sorts an adjecency list with length d in worst-case: O(d log(d))

is called for each adjecency list, i.e., self.vertex_count() times

pub fn graph(&self) -> &Graph<T>

returns reference to the underlying topology aka, the Graph<T>

Use this to call functions regarding the topology

pub fn at(&self, index: usize) -> &T

access additional information at vertex

pub fn at_mut(&mut self, index: usize) -> &mut T

mutable access of additional information at index