Struct net_ensembles::sw::SwEnsemble

pub struct SwEnsemble<T: Node, R: Rng>where
    T: Node,
    R: Rng,{ /* private fields */ }

Implements small-world graph ensemble

for more details look at documentation of module sw

Sampling

for markov steps look at MarkovChain trait

for simple sampling look at SimpleSample trait

Other

for topology functions look at GenericGraph

to access underlying topology or manipulate additional data look at WithGraph trait

to use or swap the random number generator, look at HasRng trait

Minimal example

use net_ensembles::{SwEnsemble, EmptyNode};
use net_ensembles::traits::*; // I recommend always using this
use rand_pcg::Pcg64; //or whatever you want to use as rng
use net_ensembles::rand::SeedableRng; // I use this to seed my rng, but you can use whatever

let rng = Pcg64::seed_from_u64(12);

// now create small-world ensemble with 200 nodes
// and a rewiring probability of 0.3 for each edge
let sw_ensemble = SwEnsemble::<EmptyNode, Pcg64>::new(200, 0.3, rng);

Simple sampling example

use net_ensembles::{SwEnsemble, EmptyNode};
use net_ensembles::traits::*; // I recommend always using this
use rand_pcg::Pcg64; //or whatever you want to use as rng
use net_ensembles::rand::SeedableRng; // I use this to seed my rng, but you can use whatever
use std::fs::File;
use std::io::{BufWriter, Write};

let rng = Pcg64::seed_from_u64(122);

// now create small-world ensemble with 100 nodes
// and a rewiring probability of 0.3 for each edge
let mut sw_ensemble = SwEnsemble::<EmptyNode, Pcg64>::new(100, 0.3, rng);

// setup file for writing
let f = File::create("simple_sample_sw_example.dat")
    .expect("Unable to create file");
let mut f = BufWriter::new(f);
f.write_all(b"#diameter bi_connect_max average_degree\n")
    .unwrap();

// simple sample for 10 steps
sw_ensemble.simple_sample(10,
    |ensemble|
    {
        let diameter = ensemble.graph()
            .diameter()
            .unwrap();

        let bi_connect_max = ensemble.graph()
            .clone()
            .vertex_biconnected_components(false)[0];

        let average_degree = ensemble.graph()
            .average_degree();

        write!(f, "{} {} {}\n", diameter, bi_connect_max, average_degree)
            .unwrap();
    }
);

// or just collect this into a vector to print or do whatever
let vec = sw_ensemble.simple_sample_vec(10,
    |ensemble|
    {
        let diameter = ensemble.graph()
            .diameter()
            .unwrap();

        let transitivity = ensemble.graph()
            .transitivity();
        (diameter, transitivity)
    }
);
println!("{:?}", vec);

Save and load example

only works if feature "serde_support" is enabled

Note: "serde_support" is enabled by default

I need the #[cfg(feature = "serde_support")] to ensure the example does compile if you opt out of the default feature

you can do not have to use serde_json, look here for more info

use net_ensembles::traits::*; // I recommend always using this
use serde_json;
use rand_pcg::Pcg64;
use net_ensembles::{SwEnsemble, EmptyNode, rand::SeedableRng};
use std::fs::File;

let rng = Pcg64::seed_from_u64(95);
// create small-world ensemble
let sw_ensemble = SwEnsemble::<EmptyNode, Pcg64>::new(200, 0.3, rng);

#[cfg(feature = "serde_support")]
{
    // storing the ensemble in a file:

    let sw_file = File::create("store_SW.dat")
          .expect("Unable to create file");

    // or serde_json::to_writer(sw_file, &sw_ensemble);
    serde_json::to_writer_pretty(sw_file, &sw_ensemble);

    // loading ensemble from file:

    let mut read = File::open("store_SW.dat")
        .expect("Unable to open file");

    let sw: SwEnsemble::<EmptyNode, Pcg64> = serde_json::from_reader(read).unwrap();
}

Implementations

impl<T, R> SwEnsemble<T, R>where
T: Node + SerdeStateConform,
R: Rng,

pub fn new(n: u32, r_prob: f64, rng: R) -> Self

Initialize

create new SwEnsemble graph with n vertices

r_prob is probability of rewiring for each edge

rng is consumed and used as random number generator in the following

internally uses SwGraph<T>::new(n)

pub fn draw_edge(&mut self) -> (u32, u32)

draws random edge (i0, i1)

edge rooted at i0

uniform probability

result dependent on order of adjecency lists

mut because it uses the rng

pub fn sort_adj(&mut self)

Sort adjecency lists

If you depend on the order of the adjecency lists, you can sort them

Performance

internally uses pattern-defeating quicksort as long as that is the standard

sorts an adjecency list with length d in worst-case: O(d log(d))

is called for each adjecency list, i.e., self.vertex_count() times

pub fn r_prob(&self) -> f64

returns rewiring probability

pub fn set_r_prob(&mut self, r_prob: f64)

set new value for rewiring probability

Note

will only set the value, which will be used from now on

if you also want to create a new sample, call randomize afterwards

pub fn contained_iter_neighbors_mut(
&mut self,
index: usize
) -> NContainedIterMut<'_, T, SwContainer<T>>

retuns GenericGraph::contained_iter_neighbors_mut

otherwise you would not have access to this function, since no mut access to the graph is allowed