Module net_ensembles::traits

You should use net_ensembles::traits::*

I recommend doing so for a smooth experience

contains traits you should use for accessing complete functionallity

Enums

IterWrapper

Wrapper for iterators

Traits

AdjContainer

Defines methods all adjecency containers should have such that GenericGraph can use it

Ensemble

Sampling an Ensemble

EnsembleRng

Access random number generator of Ensemble

Node

What every node should be able to do