Trait net_ensembles::traits::GraphIteratorsMut

pub trait GraphIteratorsMut<'a, T, G, A> where
    Self: WithGraph<T, G>,
    T: Node,
    A: AdjContainer<T>, {
    fn contained_iter_neighbors_mut(
        &'a mut self, 
        index: usize
    ) -> NContainedIterMut<'a, T, A>;

    fn contained_iter_mut(&'a mut self) -> ContainedIterMut<'a, T, A>;
}

Collection mut Graph iterators

Required methods

fn contained_iter_neighbors_mut(
&'a mut self,
index: usize
) -> NContainedIterMut<'a, T, A>

iterate over mutable additional information of neighbors of vertex index

iterator returns &mut T

sort_adj will affect the order

panics if index out of bounds

fn contained_iter_mut(&'a mut self) -> ContainedIterMut<'a, T, A>

get iterator over mutable additional information stored at each vertex in order of the indices

iterator returns a Node (for example EmptyNode or whatever you used)

Loading content...